move UpstreamPolicy and fix some implementation
