+++ CMakeLists_Headers.txt
@@ -84,7 +84,7 @@ set(INCLUDE_HEADERS
 	src/factory/WFOperator.h
 	src/factory/WFNameService.h
 	src/factory/WFDNSResolver.h
-	src/manager/UPSPolicy.h
+	src/manager/UpstreamPolicy.h
 )
 
 if(KAFKA STREQUAL "y")
